+++ pytest_typing_runner/errors.py
@@ -1,7 +1,7 @@
 import dataclasses
 
 
-@dataclasses.dataclass(frozen=True, kw_only=True)
+@dataclasses.dataclass(kw_only=True)
 class PyTestTypingRunnerException(Exception):
     """
     Parent exception for all exceptions from this library