+++ pytest_typing_runner/parse/errors.py
@@ -4,14 +4,14 @@ from collections.abc import Sequence
 from .. import errors
 
 
-@dataclasses.dataclass(frozen=True, kw_only=True)
+@dataclasses.dataclass(kw_only=True)
 class InvalidLine(errors.PyTestTypingRunnerException):
     """
     Base exception for problems with parsing lines
     """
 
 
-@dataclasses.dataclass(frozen=True, kw_only=True)
+@dataclasses.dataclass(kw_only=True)
 class InvalidMsgMaker(errors.PyTestTypingRunnerException):
     """
     Raised when an unknown msg maker is asked for
@@ -27,7 +27,7 @@ class InvalidMsgMaker(errors.PyTestTypingRunnerException):
             return f"Asked for a msg maker ({self.want}) but none are available to choose from"
 
 
-@dataclasses.dataclass(frozen=True, kw_only=True)
+@dataclasses.dataclass(kw_only=True)
 class InvalidInstruction(InvalidLine):
     """
     Raised when a line that looks like an invalid instruction comment is found
@@ -40,7 +40,7 @@ class InvalidInstruction(InvalidLine):
         return f"{self.reason}: {self.line}"
 
 
-@dataclasses.dataclass(frozen=True, kw_only=True)
+@dataclasses.dataclass(kw_only=True)
 class TooManyModifyLines(InvalidLine):
     """
     Raised when multiple ``modify_lines`` are produced for a single line
@@ -52,7 +52,7 @@ class TooManyModifyLines(InvalidLine):
         return f"Can only modify a line once for all collected comments: line='{self.line}'"
 
 
-@dataclasses.dataclass(frozen=True, kw_only=True)
+@dataclasses.dataclass(kw_only=True)
 class InvalidMypyOutputLine(InvalidLine):
     """
     Raised when an invalid line from mypy output is encountered
@@ -64,7 +64,7 @@ class InvalidMypyOutputLine(InvalidLine):
         return f"Line from mypy output is invalid: {self.line}"
 
 
-@dataclasses.dataclass(frozen=True, kw_only=True)
+@dataclasses.dataclass(kw_only=True)
 class UnknownSeverity(InvalidMypyOutputLine):
     """
     Raised when a severity is encountered that is unknown